+++ src/term.h
@@ -471,6 +471,8 @@ struct term {
   bool has_focus;
   bool focus_reported;
   bool in_vbell;
+  bool margin_bell;
+  bool ring_enabled;
 
   bool vt220_keys;
   bool shortcut_override;
@@ -483,6 +485,7 @@ struct term {
   bool app_keypad;
   bool auto_repeat;
   int bell_vol;
+  int margin_bell_vol;
   bool bell_taskbar; // xterm: bellIsUrgent; switchable with CSI ? 1042 h/l
   bool bell_popup;   // xterm: popOnBell;    switchable with CSI ? 1043 h/l
   bool wheel_reporting_xterm; // xterm: alternateScroll