+++ src/winpriv.h
@@ -92,6 +92,7 @@ extern void win_mouse_wheel(POINT wpos, bool horizontal, int delta);
 extern void win_mouse_move(bool nc, LPARAM);
 
 extern void win_key_reset(void);
+extern void provide_input(wchar);
 extern bool win_key_down(WPARAM, LPARAM);
 extern bool win_key_up(WPARAM, LPARAM);
 extern void do_win_key_toggle(int vk, bool on);