+++ wiki/Changelog.md
@@ -1,6 +1,7 @@
 Terminal features
   * Limit coordinates of all mouse reports to text area (#972).
   * Support bell volume escape sequence DECSWBV (#974) with wave files.
+  * Support margin bell feature (#974), with DECSET 44 enabling sequence (xterm).
 
 Character encoding
   * Enhanced and documented GB18030 support (mintty/wsltty#224).